+++ Source/Core/Credit.js
@@ -1,11 +1,13 @@
 define([
-        './defined',
-        './defineProperties',
-        './DeveloperError'
-    ], function(
-        defined,
-        defineProperties,
-        DeveloperError) {
+    './defaultValue',
+    './defined',
+    './defineProperties',
+    './DeveloperError'
+], function(
+    defaultValue,
+    defined,
+    defineProperties,
+    DeveloperError) {
     'use strict';
 
     var nextCreditId = 0;
@@ -17,6 +19,7 @@ define([
      * @param {String} [text] The text to be displayed on the screen if no imageUrl is specified.
      * @param {String} [imageUrl] The source location for an image
      * @param {String} [link] A URL location for which the credit will be hyperlinked
+     * @param {String} [showInPopup=true] If false, the credit will be visible in the main credit container.  Otherwise, it will appear in a popover
      *
      * @alias Credit
      * @constructor
@@ -25,7 +28,7 @@ define([
      * //Create a credit with a tooltip, image and link
      * var credit = new Cesium.Credit('Cesium', '/images/cesium_logo.png', 'http://cesiumjs.org/');
      */
-    function Credit(text, imageUrl, link) {
+    function Credit(text, imageUrl, link, showInPopup) {
         var hasLink = (defined(link));
         var hasImage = (defined(imageUrl));
         var hasText = (defined(text));
@@ -45,6 +48,7 @@ define([
         this._link = link;
         this._hasLink = hasLink;
         this._hasImage = hasImage;
+        this._showInPopup = defaultValue(showInPopup, true);
 
         // Credits are immutable so generate an id to use to optimize equal()
         var id;
@@ -107,6 +111,18 @@ define([
         id : {
             get : function() {
                 return this._id;
+            }
+        },
+
+        /**
+         * Whether the credit should be displayed on screen or in a lightbox
+         * @memberof Credit.prototype
+         * @type {Boolean}
+         * @readonly
+         */
+        showInPopup : {
+            get : function() {
+                return this._showInPopup;
             }
         }
     });